requirements.txt
streamlit>=1.27.0
Pillow>=10.0.0
pandas>=2.0.0
boto3>=1.28.0
python-dotenv>=1.0.0
xlsxwriter>=3.0.0 